+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization/Http/IEndpointAuthorizationLogger.cs
@@ -12,21 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Text.Json.Serialization;
+using Energinet.DataHub.MarketParticipant.Authorization.Model;
+using Energinet.DataHub.MarketParticipant.Authorization.Model.AccessValidationRequests;
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Model;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Http;
 
-public sealed class Signature
+public interface IEndpointAuthorizationLogger
 {
-    [JsonPropertyName("signature")]
-    public required string Value { get; init; }
-
-    [JsonPropertyName("expires")]
-    public required DateTimeOffset Expires { get; init; }
-
-    [JsonPropertyName("keyVersion")]
-    public required string KeyVersion { get; init; }
-
-    [JsonPropertyName("requestId")]
-    public required Guid RequestId { get; init; }
+    Task LogAsync(ILoggableAccessRequest accessRequest, AuthorizationResult authorizationResult);
 }